// sdk/ai/azure-ai-client/src/models/models.ts
export interface Evaluation {
  /** Identifier of the evaluation. */
  id?: string;
  /** Data for evaluation. */
  data: InputDataUnion;
  /** Display Name for evaluation. It helps to find evaluation easily in AI Studio. It does not need to be unique. */
  displayName?: string;
  /** Description of the evaluation. It can be used to store additional information about the evaluation and is mutable. */
  description?: string;
  /** Metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
  /** Status of the evaluation. It is set by service and is read-only. */
  readonly status?: string;
  /** Evaluation's tags. Unlike properties, tags are fully mutable. */
  tags?: Record<string, string>;
  /** Evaluation's properties. Unlike tags, properties are add-only. Once added, a property cannot be removed. */
  properties?: Record<string, string>;
  /** Evaluators to be used for the evaluation. */
  evaluators: Record<string, EvaluatorConfiguration>;
}
// ...
export interface InputData {
  /** the discriminator possible values: app_insights, dataset */
  type: string;
}
// ...
export interface AppInsightsConfiguration extends InputData {
  readonly type: "app_insights";
  /** LogAnalytic Workspace resourceID associated with AppInsights */
  resourceId: string;
  /** Query to fetch the data. */
  query: string;
  /** Service name. */
  serviceName: string;
}

export function appInsightsConfigurationSerializer(
  item: AppInsightsConfiguration,
): Record<string, unknown> {
  return {
    resourceId: item["resourceId"],
    query: item["query"],
    serviceName: item["serviceName"],
  };
}

/** Dataset as source for evaluation. */
export interface Dataset extends InputData {
  readonly type: "dataset";
  /** Evaluation input data */
  id: string;
}

export function datasetSerializer(item: Dataset): Record<string, unknown> {
  return {
    Uri: item["id"],
  };
}

/** Metadata pertaining to creation and last modification of the resource. */
export interface SystemData {
  /** The timestamp the resource was created at. */
  readonly createdAt?: Date;
  /** The identity that created the resource. */
  readonly createdBy?: string;
  /** The identity type that created the resource. */
  readonly createdByType?: string;
  /** The timestamp of resource last modification (UTC) */
  readonly lastModifiedAt?: Date;
}

/** Evaluator Configuration */
export interface EvaluatorConfiguration {
  /** Identifier of the evaluator. */
  id: string;
  /** Initialization parameters of the evaluator. */
  initParams?: Record<string, any>;
  /** Data parameters of the evaluator. */
  dataMapping?: Record<string, string>;
}
// ...
export interface _PagedEvaluation {
  /** The list of Evaluations. */
  value: Evaluation[];
  /** The link to the next page of items */
  nextLink?: string;
}

/** Update Evaluation Request */
export interface UpdateEvaluationRequest {
  /** Tags to be updated. */
  tags: Record<string, string>;
  /** Display Name */
  displayName: string;
  /** Description */
  description: string;
}

export function updateEvaluationRequestSerializer(
  item: UpdateEvaluationRequest,
): Record<string, unknown> {
  return {
    tags: serializeRecord(item.tags as any) as any,
    displayName: item["displayName"],
    description: item["description"],
  };
}

/** Evaluation Schedule Definition */
export interface EvaluationSchedule {
  /** Identifier of the evaluation. */
  id?: string;
  /** Data for evaluation. */
  data: InputDataUnion;
  /** Display Name for evaluation. It helps to find evaluation easily in AI Studio. It does not need to be unique. */
  displayName?: string;
  /** Description of the evaluation. It can be used to store additional information about the evaluation and is mutable. */
  description?: string;
  /** Metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
  /** Status of the evaluation. It is set by service and is read-only. */
  readonly status?: string;
  /** Evaluation's tags. Unlike properties, tags are fully mutable. */
  tags?: Record<string, string>;
  /** Evaluation's properties. Unlike tags, properties are add-only. Once added, a property cannot be removed. */
  properties?: Record<string, string>;
  /** Evaluators to be used for the evaluation. */
  evaluators: Record<string, EvaluatorConfiguration>;
  /** Recurrence pattern for the evaluation. */
  recurrence?: Recurrence;
  /** Cron expression for the evaluation. */
  cronExpression?: string;
  /** Sampling strategy for the evaluation. */
  samplingStrategy: SamplingStrategy;
}
// ...
export interface Recurrence {
  /** The frequency to trigger schedule. */
  frequency: Frequency;
  /** Specifies schedule interval in conjunction with frequency */
  interval: number;
  /** The recurrence schedule. */
  schedule: RecurrenceSchedule;
}

export function recurrenceSerializer(
  item: Recurrence,
): Record<string, unknown> {
  return {
    frequency: item["frequency"],
    interval: item["interval"],
    schedule: recurrenceScheduleSerializer(item.schedule),
  };
}

/** Frequency of the schedule - day, week, month, hour, minute */
export type Frequency = "Month" | "Week" | "Day" | "Hour" | "Minute";

/** RecurrenceSchedule Definition */
export interface RecurrenceSchedule {
  /** List of hours for the schedule. */
  hours: number[];
  /** List of minutes for the schedule. */
  minutes: number[];
  /** List of days for the schedule. */
  weekDays: WeekDays[];
  /** List of month days for the schedule */
  monthDays: number[];
}
// ...
export interface SamplingStrategy {
  /** Sampling rate */
  rate: number;
}

export function samplingStrategySerializer(
  item: SamplingStrategy,
): Record<string, unknown> {
  return {
    rate: item["rate"],
  };
}

/** Paged evaluation schedule items */
export interface _PagedEvaluationSchedule {
  /** The list of Evaluation Schedules. */
  value: EvaluationSchedule[];
  /** The link to the next page of items */
  nextLink?: string;
}
// ...
export type InputDataUnion = AppInsightsConfiguration | Dataset | InputData;