// sdk/ai/azure-ai-client/src/models/options.ts
export interface EvaluationsCreateOptionalParams extends OperationOptions {}

/** Optional parameters. */
export interface EvaluationsListOptionalParams extends OperationOptions {
  /** The number of result items to return. */
  top?: number;
  /** The number of result items to skip. */
  skip?: number;
  /** The maximum number of result items per page. */
  maxpagesize?: number;
}

/** Optional parameters. */
export interface EvaluationsUpdateOptionalParams extends OperationOptions {}

/** Optional parameters. */
export interface EvaluationsGetOptionalParams extends OperationOptions {}

/** Optional parameters. */
export interface EvaluationsCreateScheduleOptionalParams
  extends OperationOptions {}

/** Optional parameters. */
export interface EvaluationsGetScheduleOptionalParams
  extends OperationOptions {}

/** Optional parameters. */
export interface EvaluationsListSchedulesOptionalParams
  extends OperationOptions {
  /** The number of result items to return. */
  top?: number;
  /** The number of result items to skip. */
  skip?: number;
  /** The maximum number of result items per page. */
  maxpagesize?: number;
}

/** Optional parameters. */
export interface EvaluationsListScheduleEvaluationsOptionalParams
  extends OperationOptions {
  /** The number of result items to return. */
  top?: number;
  /** The number of result items to skip. */
  skip?: number;
  /** The maximum number of result items per page. */
  maxpagesize?: number;
}

/** Optional parameters. */
export interface EvaluationsDeleteScheduleOptionalParams
  extends OperationOptions {}